requirements.txt
matplotlib
numpy
mpl-interactions
Pillow
